baymax-backend/requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv>=1.0.1,<2.0.0
pydantic==2.7.4
httpx>=0.27.0
python-multipart==0.0.22

# Database
asyncpg==0.29.0

# LangChain Ecosystem
langchain==0.3.0
langchain-groq==0.2.0
langgraph==0.5.4

# Embeddings
sentence-transformers==3.0.0
torch==2.10.0

# Vector DB
pinecone-client==5.0.0
pinecone-text==0.9.0

# Cache
redis==5.0.0

# Web Search
ddgs>=9.0.0

# Auth
PyJWT>=2.8.0
bcrypt>=4.1.0